String parsers moved to a separate file
